refactor(reporter): drop unused cross-run recommendation aggregation
The aggregate fold was exported and tested but never wired into renderHtml
or the report template, so it shipped as dead code. The HTML report renders
recommendations per run only. Remove the aggregation and keep the per-run
panel tests in a focused file.
